requirements.txt
# local package
-e .

# external requirements
click
coverage
awscli
flake8
python-dotenv>=0.5.1
seaborn
numpy
datasets
dash
transformers
scikit-learn
torch
sagemaker


